backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
torch
torchvision
matplotlib
pillow
scikit-image
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
